Update README.md
This project required me to write C++ code that incorporates prewritten Java code. The requirements were to create a user menu interface that could be used to add, store, load, and edit animal data. I believe this project showcases my ability to write code that is easy to follow and understand and functions efficiently. I could enhance the code by adding more classes for the specific animal species in order to include more detailed data that is personalized. Incorporating C++ with Java was the most difficult part of this project, but having access to the Hello World assignment we completed that introduced us to that practice was very helpful. Having this knowledge will be helpful in the future because I will likely do much more with combining programming languages to enhance certain types of programs.

M. Duprey